BAsic Setup
This is a basic working ASM setup. I may want to refer to this in the
future, as it has everything in it that I know what it does, and lacks the
things I have no clue about yet, such as BASIC and FARSTACK in the MODEL.

THis assembles. I'll start including the random stuff I see in my
professor's boilerplate when I understand what it does. This makes me feel
more in control of what's going on. Perhaps at some point I'll run into an
issue where the omissions cause problems. So far, so good, however.