BITCOUNT
Bit count proc working. Pass a value via the stack, and the number of set
bits is returned via AL.

Now to Mask for even or odd detection.

Also, an asm discovery: in order to implicitly pull a value from the stack,
the MODEL must be set up with either C or BASIC setting. I know the C setting
sets up procs for inclusion as C modules, so I am assuming the BASIC sets up
for inclusion as a module for the BASIC language.

The actual information on this is supprisingly hard to find. I learned about
the MODEL C op from a library book. The internet and most books only address
the memory size parameter of MODEL. I don't understand why when presenting
MODEL, all parameters are not just given to you.

I plan to track down a difinitive spec sheet for this language.
Instructional texts are too affraid to overwhelm me with too much
information. THis approach does not sit with me.