Encoding
Just the design on this push.

Now that I have my feet wet in asm, the logic of the projects is becommin
gmore and more complex. I find I need to get back to a proper programming
approach: start with a design, code in short modules, unit test. This is
going to save me a lot of headaches.

For this commit, all I have is the design in a text file. Nothing fancy,
just some asm dialect of pseudocode. This should be a snap to convert to
code and test, module by module.