Encode start
Starting with another raw, basic file. This one has some tests in it. I
wanted to explicitly assign the DS reg to the data segment. With just an
unlabled .DATA, it seemed that the ES took the data address. I also couldnt
add a lable to .DATA without an error in the assembler.

.FARDATA allows me to lable the data seg and use ASSUME to point to it. I'm
not sure of the why though, so some research is needed.