Encode: Setup
Setup (needs comment headder) proc added and tested in driver. Breaks up the
provided bits into encoding groups and rotates them into positions reserved
for data bits in the final encoded valuses.

Bits 1,2,4,and 8 will be parity bits. The other 7 bits masked and assembled
in this proc are reserved for the data to be encoded in this 11 bit encode.