Encoding: report
REPORT subproc was added, and MAIN was constructed. There were problems:

It turns out that using .FARDATA and assigning it to DS caused problems.
Loading the address of the string locations did not work with LEA, so the
output strings produced garbage.

I went back and restored the memory seg to .DATA, which does not take a
label. The ASSUME command was therefore also removed, and the data segment
was set up to point to ES. This setup allows LEA to get the correct offset
to show the strings properly.

However, now the entered and encoded values are returning unexpected
results.

Going to see if this is a REPORT output issue or if the data seg alterations
destroyed my logic after a break.